commonlib/storage: Zero extend MMC capacity
Fix CID 1376472 detected by coverity.  Zero extend the capacity instead
of sign extending it.

TEST=Build and run on reef

Reviewed-on: https://review.coreboot.org/20392
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/commonlib/storage/mmc.c
@@ -433,7 +433,7 @@ int mmc_update_capacity(struct storage_media *media)
 	 * ext_csd's capacity is valid if the value is
 	 * more than 2GB
 	 */
-	capacity = (ext_csd[EXT_CSD_SEC_CNT + 0] << 0 |
+	capacity = (uint32_t)(ext_csd[EXT_CSD_SEC_CNT + 0] << 0 |
 		    ext_csd[EXT_CSD_SEC_CNT + 1] << 8 |
 		    ext_csd[EXT_CSD_SEC_CNT + 2] << 16 |
 		    ext_csd[EXT_CSD_SEC_CNT + 3] << 24);
